Moved PubChemElements_all.json back
