+++ src/main/resources/graphics/scenery/PubChemElements_all.json
